Move api request to a separate file
