+++ src/rvt.js
@@ -1,7 +1,6 @@
 import { getUUID } from "./getUUID.js";
 import { getSignature } from "./getSignature.js";
 import { yandexProtobuf } from "./yandexProtobuf.js";
-import { workerHost, yandexUserAgent } from "./config/config.js";
 import debug from "./utils/debug.js";
 
 // Request video translation from Yandex API
@@ -12,54 +11,34 @@ async function requestVideoTranslation(
   responseLang,
   callback
 ) {
-  // Initialize variables
-  const deviceId = getUUID(true);
-  const body = yandexProtobuf.encodeTranslationRequest(
-    url,
-    deviceId,
-    duration,
-    requestLang,
-    responseLang
-  );
-
   try {
     debug.log("requestVideoTranslation");
-    // Create a fetch options object with headers and body
-    const options = {
-      // url: `https://${workerHost}/stream-translation/whitelist-stream`,
-      // url: `https://${workerHost}/stream-translation/translate-stream`,
-      url: `https://${workerHost}/video-translation/translate`,
-      method: "POST",
-      headers: {
-        Accept: "application/x-protobuf",
-        "Accept-Language": "en",
-        "Content-Type": "application/x-protobuf",
-        "User-Agent": yandexUserAgent,
-        Pragma: "no-cache",
-        "Cache-Control": "no-cache",
-        "Sec-Fetch-Mode": "no-cors",
-        "sec-ch-ua": null,
-        "sec-ch-ua-mobile": null,
-        "sec-ch-ua-platform": null,
+    const yar = await import(
+      `./yandexRequest${BUILD_MODE === "cloudflare" ? "-cloudflare" : ""}.js`
+    );
+    const yandexRequest = yar.default;
+    debug.log("Inited yandexRequest...");
+    // Initialize variables
+    const deviceId = getUUID(true);
+    const body = yandexProtobuf.encodeTranslationRequest(
+      url,
+      deviceId,
+      duration,
+      requestLang,
+      responseLang
+    );
+    // Send the request
+    await yandexRequest(
+      // "/stream-translation/whitelist-stream",
+      // "/stream-translation/translate-stream",
+      "/video-translation/translate",
+      body,
+      {
         "Vtrans-Signature": await getSignature(body),
         "Sec-Vtrans-Token": getUUID(false),
       },
-      binary: true,
-      data: new Blob([body]),
-      responseType: "arraybuffer",
-    };
-    // Send the request using GM_xmlhttpRequest
-    GM_xmlhttpRequest({
-      ...options,
-      onload: (http) => {
-        debug.log(http.status, http);
-        callback(http.status === 200, http.response);
-      },
-      onerror: (error) => {
-        console.error("[VOT]", error);
-        callback(false);
-      },
-    });
+      callback,
+    );
   } catch (exception) {
     console.error("[VOT]", exception);
     // Handle errors
@@ -68,3 +47,5 @@ async function requestVideoTranslation(
 }
 
 export default requestVideoTranslation;
+
+unsafeWindow.requestVideoTranslation = requestVideoTranslation;
