+++ src/yandexRequest-cloudflare.js
@@ -1,32 +1,17 @@
-import { getUUID } from "./getUUID.js";
-import { getSignature } from "./getSignature.js";
-import { yandexProtobuf } from "./yandexProtobuf.js";
 import { workerHost } from "./config/config-cloudflare.js";
 import { yandexUserAgent } from "./config/config.js";
 import debug from "./utils/debug.js";
 
-// Request video translation from Yandex API
-async function requestVideoTranslation(
-  url,
-  duration,
-  requestLang,
-  responseLang,
+async function yandexRequest(
+  path,
+  body,
+  headers,
   callback
 ) {
-  // Initialize variables
   let response;
   let responseBody;
-  const deviceId = getUUID(true);
-  const body = yandexProtobuf.encodeTranslationRequest(
-    url,
-    deviceId,
-    duration,
-    requestLang,
-    responseLang
-  );
-
   try {
-    debug.log("requestVideoTranslation");
+    debug.log("yandexRequest:", path);
     // Create a fetch options object with headers and body
     const options = {
       method: "POST",
@@ -39,15 +24,16 @@ async function requestVideoTranslation(
       referrerPolicy: "no-referrer",
       body: JSON.stringify({
         headers: {
-          Accept: "application/x-protobuf",
-          "Accept-Language": "en",
-          "Content-Type": "application/x-protobuf",
-          "User-Agent": yandexUserAgent,
-          Pragma: "no-cache",
-          "Cache-Control": "no-cache",
-          "Sec-Fetch-Mode": "no-cors",
-          "Vtrans-Signature": await getSignature(body),
-          "Sec-Vtrans-Token": getUUID(false),
+          ...{
+            "Accept": "application/x-protobuf",
+            "Accept-Language": "en",
+            "Content-Type": "application/x-protobuf",
+            "User-Agent": yandexUserAgent,
+            "Pragma": "no-cache",
+            "Cache-Control": "no-cache",
+            "Sec-Fetch-Mode": "no-cors",
+          },
+          ...headers
         },
         body: Array.from(body)
       }),
@@ -57,7 +43,7 @@ async function requestVideoTranslation(
       `https://${workerHost}/video-translation/translate`,
       options
     );
-    debug.log(response.status, response);
+    debug.log("yandexRequest:", response.status, response);
     // Get the response body as an array buffer
     responseBody = await response.arrayBuffer();
   } catch (exception) {
@@ -66,9 +52,11 @@ async function requestVideoTranslation(
     response = { status: -1 };
     responseBody = exception;
   }
-
+  
   // Call the callback function with the result
   callback(response.status == 200, responseBody);
 }
 
-export default requestVideoTranslation;
+export default yandexRequest;
+
+unsafeWindow.yandexRequest = yandexRequest;
